Struct libp2p_floodsub::Floodsub

pub struct Floodsub { /* fields omitted */ }

Network behaviour that handles the floodsub protocol.

Implementations

impl Floodsub

pub fn new(local_peer_id: PeerId) -> Self

Creates a Floodsub with default configuration.

pub fn from_config(config: FloodsubConfig) -> Self

Creates a Floodsub with the given configuration.

pub fn add_node_to_partial_view(&mut self, peer_id: PeerId)

Add a node to the list of nodes to propagate messages to.

pub fn remove_node_from_partial_view(&mut self, peer_id: &PeerId)

Remove a node from the list of nodes to propagate messages to.

pub fn subscribe(&mut self, topic: Topic) -> bool

Subscribes to a topic.

Returns true if the subscription worked. Returns false if we were already subscribed.

pub fn unsubscribe(&mut self, topic: Topic) -> bool

Unsubscribes from a topic.

Note that this only requires the topic name.

Returns true if we were subscribed to this topic.

pub fn publish(&mut self, topic: impl Into<Topic>, data: impl Into<Vec<u8>>)

Publishes a message to the network, if we’re subscribed to the topic only.

pub fn publish_any(&mut self, topic: impl Into<Topic>, data: impl Into<Vec<u8>>)

Publishes a message to the network, even if we’re not subscribed to the topic.

pub fn publish_many(
&mut self,
topic: impl IntoIterator<Item = impl Into<Topic>>,
data: impl Into<Vec<u8>>
)

Publishes a message with multiple topics to the network.

Note: Doesn’t do anything if we’re not subscribed to any of the topics.

pub fn publish_many_any(
&mut self,
topic: impl IntoIterator<Item = impl Into<Topic>>,
data: impl Into<Vec<u8>>
)

Publishes a message with multiple topics to the network, even if we’re not subscribed to any of the topics.